Added initial tracking of history of values. Works with the concept of displayed values, though no record is kept of displayed values at this stage. Displayed values are identified by address and time. SinglePrintable instances keep the set of displayed values that have contributed to the value they currently hold. When a printable is displayed, this history is replaced with a single item, the new displayed value created for that display.
Only two operation types are recorded for values;
assignment, e.g. a = b  the history of a is replaced by the history of b.
modification, e.g. a += b  the history of a is replaced with the union of the histories of a and b, as both values were used

Things like a = b + c are implemented as:
assign b to a
modify a using c

git-svn-id: https://svn.ece.auckland.ac.nz/svn/p4p-2010-g79@11 a75856b6-468a-0410-9532-ba8c368bb4fd