* Fixing synchronisation bug:
Problem: When an action is lodged in the Displayer (on the main thread) before the Displayer render loop (on the render thread) is about to call World::AcquireReaderLock() to display Viewables (without animation). AcquireReaderLock() waits for all pending actions to complete before granting this lock. The Displayer cannot animate/perform/complete the pending action until the next iteration of the render loop which it never reaches due to AcquireReaderLock blocking

Solution: Add an argument to World::AcquireReaderLock allowing a VO read lock to be acquired even if there is an action pending. Intended to be used only by the Displayer at this stage.

* Adding a bit of debugging info for actions.

* Cleaning up, doing TODOs etc.


git-svn-id: https://svn.ece.auckland.ac.nz/svn/p4p-2010-g79@23 a75856b6-468a-0410-9532-ba8c368bb4fd