Implemented action buffer. This buffer will later be used for collapsing actions that should be performed together.
To implement the buffer properly, a bunch of logic was moved out of registry and into the actual actions. This was necessary because things like Register only occur when an action completes, possibly long after registry is notified of some related assignment, so registry can't go ahead and grab the representation or anything. This makes the suppress functions (and perhaps other stuff) fairly pointless, but they have not been removed yet. I've also started to wonder if registry is performing two roles and should be split.

git-svn-id: https://svn.ece.auckland.ac.nz/svn/p4p-2010-g79@31 a75856b6-468a-0410-9532-ba8c368bb4fd