SinglePrintablle hackery hopefully mostly gone, however, still not that nice as collections are often treated differently to items, be they SinglePrintable or otherwise.
There is now a factory stored in the history for each element, such that viewables with the appropriate parameters can be created on demand. This means that the actual creation is still happening on add to array, so this action will need a special undo.

Matrix is broken but I didn't check it was working before, let me know.

git-svn-id: https://svn.ece.auckland.ac.nz/svn/p4p-2010-g79@70 a75856b6-468a-0410-9532-ba8c368bb4fd